requirements.txt
aioredis==2.0.1
alembic==1.8.1
anyio==3.5.0
arrow==1.2.2
asgiref==3.5.0
astroid==2.9.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
bcrypt==3.2.0
bidict==0.22.1
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==1.7.3
cryptography==36.0.1
ecdsa==0.17.0
exceptiongroup==1.0.4
fastapi==0.74.1
fastapi-utils==0.2.1
greenlet==2.0.1
h11==0.13.0
httpcore==0.16.2
httptools==0.3.0
httpx==0.23.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.0.3
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
Mako==1.2.4
MarkupSafe==2.1.0
mccabe==0.6.1
minio==7.1.13
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pylint==2.12.2
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.4
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.0
python-socketio==5.7.2
PyYAML==6.0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.44
starlette==0.17.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
wrapt==1.13.3
redis==4.2.2